requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-import-export==3.3.5
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
selenium==4.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tablib==3.5.0
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0